// src/generated/models/accountDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Carmentis indexer API
 * Documentation for the Carmentis indexer API.
 * OpenAPI spec version: 1.0
 */
import type { EscrowLockDto } from './escrowLockDto';
import type { StakingLockDto } from './stakingLockDto';
import type { VestingLockDto } from './vestingLockDto';

export interface AccountDto {
  id: string;
  publicKey: string;
  height: number;
  balance: number;
  spendable: number;
  lockedInStaking: number;
  lockedInVesting: number;
  lockedInEscrows: number;
  stakingLocks: StakingLockDto[];
  vestingLocks: VestingLockDto[];
  escrowLocks: EscrowLockDto[];
}